Split into renderer and example
